requirements.txt
asgiref==3.3.4
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
distro==1.5.0
Django==3.2.4
djangorestframework==3.12.4
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
gunicorn==20.1.0
idna==2.10
jsonschema==3.2.0
paramiko==2.7.2
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
texttable==1.6.3
urllib3==1.26.5
websocket-client==0.59.0
whitenoise==5.2.0
wincertstore==0.2
